pyproject.toml
[project]
name = "cleandesk"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.2.0",
    "plyer>=2.1.0",
    "pydantic>=2.12.5",
    "pystray>=0.19.5",
    "pytest>=9.0.2",
    "rich>=14.3.3",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
build = [
    "pyinstaller>=6.11.0",
]
